# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The lens that shapes how the eight perspectives are chosen. "wissenschaft" draws from scientific disciplines and the liberal arts, "stakeholder" from the actors involved in the topic, and "politische_debatte" from the constituent positions of the political debate.
 */
export type PerspectiveMode =
  (typeof PerspectiveMode)[keyof typeof PerspectiveMode];

export const PerspectiveMode = {
  wissenschaft: "wissenschaft",
  stakeholder: "stakeholder",
  politische_debatte: "politische_debatte",
} as const;

export interface TextPerspectiveRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  topic: string;
  mode?: PerspectiveMode;
}

export interface ImagePerspectiveRequest {
  /**
   * Base64-encoded image data (without the data URL prefix).
   * @minLength 1
   */
  image: string;
  /**
   * MIME type of the image, e.g. image/jpeg.
   * @minLength 1
   */
  mimeType: string;
  mode?: PerspectiveMode;
}

export interface Perspective {
  id: string;
  category: string;
  title: string;
  color: string;
  summary: string;
  intro: string;
  analysis: string;
  comprehensive: string;
}

export interface PerspectiveResult {
  /** The topic explored. For image requests, this is the detected object name. */
  topic: string;
  perspectives: Perspective[];
}