# lib/api-zod/src/generated/types/imagePerspectiveRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PerspectiveMode } from "./perspectiveMode";

export interface ImagePerspectiveRequest {
  /**
   * Base64-encoded image data (without the data URL prefix).
   * @minLength 1
   */
  image: string;
  /**
   * MIME type of the image, e.g. image/jpeg.
   * @minLength 1
   */
  mimeType: string;
  mode?: PerspectiveMode;
}