# lib/api-zod/src/generated/types/perspective.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Perspective {
  id: string;
  category: string;
  title: string;
  color: string;
  summary: string;
  intro: string;
  analysis: string;
  comprehensive: string;
}