# lib/api-zod/src/generated/types/perspectiveMode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * The lens that shapes how the eight perspectives are chosen. "wissenschaft" draws from scientific disciplines and the liberal arts, "stakeholder" from the actors involved in the topic, and "politische_debatte" from the constituent positions of the political debate.
 */
export type PerspectiveMode =
  (typeof PerspectiveMode)[keyof typeof PerspectiveMode];

export const PerspectiveMode = {
  wissenschaft: "wissenschaft",
  stakeholder: "stakeholder",
  politische_debatte: "politische_debatte",
} as const;