# lib/api-zod/src/generated/types/perspectiveResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Perspective } from "./perspective";

export interface PerspectiveResult {
  /** The topic explored. For image requests, this is the detected object name. */
  topic: string;
  perspectives: Perspective[];
}