# lib/api-zod/src/generated/types/textPerspectiveRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PerspectiveMode } from "./perspectiveMode";

export interface TextPerspectiveRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  topic: string;
  mode?: PerspectiveMode;
}